+++ perf/_compare.py
@@ -220,6 +220,8 @@ def bench_sort_key(item):
 
 
 def compare_suites(benchmarks, sort_benchmarks, by_speed, args):
+    # FIXME: don't group by name to not loose order
+    # kw1, kw5, kw10 must not become kw1, kw10, kw5
     grouped_by_name = benchmarks.group_by_name()
     if not grouped_by_name:
         print("ERROR: Benchmark suites have no benchmark in common",